Migrate articles page to be the homepage
