--- pages/articles.tsx
@@ -1,46 +1,0 @@
-import Link from 'next/link';
-import { GetStaticProps } from 'next';
-import Layout from '../components/layout';
-import CustomDate from '../components/custom-date';
-import { getSortedPostsData } from '../lib/posts.service';
-import utilStyles from '../styles/utils.module.scss';
-import { MongoBlogPost } from '../lib/types';
-
-export default function Articles({ allPostsData }:
-  { allPostsData: MongoBlogPost[] }
-)  {
-  return (
-    <Layout>
-      <section className={`${utilStyles.headingMd} ${utilStyles.padding1px}`}>
-          <h2 className={utilStyles.headingLg}>Blog</h2>
-          <ul className={utilStyles.list}>
-            {allPostsData.map(({ blogID, pageKey, date, title, author, hidden }) => !hidden ? (
-              <li className={utilStyles.listItem} key={parseInt(blogID)}>
-                <Link href={`/posts/${pageKey}`}>
-                  <a>{title}</a>
-                </Link>
-                <br />
-                <small>
-                  {author}&nbsp;&nbsp;/&nbsp;&nbsp;<CustomDate ms={date} />
-                </small>
-              </li>
-            ) : null )}
-          </ul>
-      </section>
-    </Layout>
-  );
-}
-
-// Get static props will get the blog posts on static generation pre-render
-export const getStaticProps: GetStaticProps = async () => {
-  // get all the Post Datas
-  const allPostsData = await getSortedPostsData();
-
-  return {
-    props: {
-      allPostsData
-    },
-    // will revalidate changes every 5 seconds
-    revalidate: 3
-  }
-}
